user機能におけるメソッドを切り分けるためのmoduleの導入
