+++ app/controllers/users/sessions_controller.rb
@@ -1,6 +1,6 @@
 class Users::SessionsController < Devise::SessionsController
   def new_guest
-    user = User.guest
+    user = LoginGuestUserService.guest
     sign_in user
     redirect_to root_path
     flash[:notice] = 'ゲストユーザーとしてログインしました。'
