+++ app/services/login_guest_user_service.rb
@@ -1,0 +1,8 @@
+class LoginGuestUserService
+  def self.guest
+    User.find_or_create_by!(email: 'guest@example.com') do |user|
+      user.password = SecureRandom.urlsafe_base64
+      user.name = 'ゲスト'
+    end
+  end
+end
