+++ app/services/update_without_current_password_service.rb
@@ -1,0 +1,15 @@
+module UpdateWithoutCurrentPasswordService
+  def update_without_current_password(params, *options)
+    params.delete(:current_password)
+
+    if params[:password].blank? && params[:password_confirmation].blank?
+
+      params.delete(:password)
+      params.delete(:password_confirmation)
+    end
+
+    result = update_attributes(params, *options)
+    clean_up_passwords
+    result
+  end
+end
